requirements.txt
tensorflow==2.15.0
tensorflow-hub==0.15.0
pillow==10.4.0
keras==2.15.0
streamlit==1.41.1
numpy==1.26.4
# tensorflow
# tensorflow-hub
# pillow
# streamlit
# numpy
